Refactor Console filtering and searching hooks to avoid redundant work
